requirements.txt
# Core
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyyaml>=6.0

# ML Models
xgboost>=2.0
prophet>=1.1
statsmodels>=0.14

# API
fastapi>=0.110
uvicorn[standard]>=0.29
jinja2>=3.1
python-multipart>=0.0.9

# MLOps
dvc>=3.40
